LayoutTests/fast/forms/color/input-color-onchange-event-expected.txt

Test if change event fires properly when color chooser changes. Bug 66848 
To manually test this, click on the input color element in the top left corner and change the value from the color chooser. See if the number of "value changed" messages matches the number of times you changed the color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS onChangeCount is 0
FAIL onChangeCount should be 1. Was 0.
FAIL onChangeCount should be 1. Was 0.
FAIL input.value should be #ff0000. Was #000000.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
